// packages/labextension/src/model/lecture.ts
/**
 * Grader Extension API Schemas
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface Lecture {
  id?: number;
  name?: string;
  code?: string;
  complete?: boolean;
}